+++ samples/android/src/main/java/ffc/genogram/android/sample/SampleActivity.kt
@@ -30,7 +30,7 @@ import ffc.genogram.Family
 import ffc.genogram.Person
 import ffc.genogram.android.Families
 import ffc.genogram.android.GenogramFragment
-import ffc.genogram.android.GenogramView
+import ffc.genogram.android.PersonViewHolder
 
 class SampleActivity : AppCompatActivity() {
 
@@ -45,14 +45,14 @@ class SampleActivity : AppCompatActivity() {
         supportFragmentManager.beginTransaction().replace(R.id.container, genogram).commit()
     }
 
-    val personVh = object : GenogramView.PersonViewHolder {
+    val personVh = object : PersonViewHolder {
 
         override fun viewFor(person: Person, context: Context, parent: ViewGroup): View {
             val view = LayoutInflater.from(context).inflate(R.layout.node_item, parent, false)
             val icon = view.findViewById<Button>(R.id.icon)
             when (person.gender) {
-                1 -> icon.setBackgroundResource(R.drawable.male_node_icon)
-                else -> icon.setBackgroundResource(R.drawable.female_node_icon)
+                0 -> icon.setBackgroundResource(R.drawable.male_node_icon)
+                1 -> icon.setBackgroundResource(R.drawable.female_node_icon)
             }
             val name = view.findViewById<TextView>(R.id.name)
             name.text = person.firstname
